fix encode_set_command struct packing
the struct packing blows up because the format string misses the `B` that encodes
the command integer.

the error was `struct.error: pack expected 20 items for packing (got 21)`

I'm not a python person so I'm not sure if I need to provide any additional info for this.
having said that, applying this patch enables me to run the .set() method on Fridge against
a `pluginfestivals icecube duo`